refactor: 패키지 분리 및 커스텀 예외 네이밍 변경
