--- src/main/java/roomescape/common/exception/custom/PaymentBadRequestException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception.custom;
-
-public class PaymentBadRequestException extends RuntimeException{
-
-    public PaymentBadRequestException(final String message) {
-        super(message);
-    }
-}
