+++ src/main/java/roomescape/common/exception/custom/PaymentClientException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.custom;
+
+public class PaymentClientException extends RuntimeException {
+
+    public PaymentClientException(final String message) {
+        super(message);
+    }
+}
